Remove the sequence number from the API
The intial idea was to let callers decide how the sequence number is stored.
This came at the expense of polluting the API with the sequence number: store
and load functions took the sequence number by (const-)reference.

With this commit, the sequence number is stored in the SeqLock itself. This leads
to a cleaner API. Callers can initialize and share the sequence number through the use
of placement new when instantiating the SeqLock between processes.