More robust util::SharedMemory
Basically the Map<T, Args...> call happens in the Factory method.
Chose that over a constructor to rely on std::expected which is neater.
This means construction and memory mapping happen in one function,
which also allows us to properly lock the file if we're the creator
and the type T is constructed in place there.